refactor: rename file
Rename file to match the name of primary type exported from it.
