+++ src/utils/env.ts
@@ -8,7 +8,7 @@ import { coerceRegistryUrl, registryUrl } from "../types/registry-url";
 import { tryGetAuthForRegistry } from "../types/upm-config";
 import { CmdOptions } from "../types/options";
 import { manifestPathFor } from "../types/project-manifest";
-import { Registry } from "../registry-client";
+import { Registry } from "../npm-client";
 
 export type Env = Readonly<{
   cwd: string;
